LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/intrinsic-size/row-005-expected.txt


FAIL .floating-flexbox 1 assert_equals:
<div class="floating-flexbox" data-expected-width="300">
      <!-- min contribution: 100 -->
      <!-- fraction: -0.5 -->
      <!-- flex base size + product: 200px + -0.5*200px = 100px -->
      <div style="flex: 1 1 200px; width:50px;">
        <div></div>
      </div>
      <!-- min contribution: 100 -->
      <!-- fraction: -0.75 -->
      <!-- flex base size + product: 400px + -0.5*400px = 200px -->
      <div style="flex: 1 1 400px; width:50px;">
        <div></div>
      </div>
    </div>
width expected 300 but got 100
FAIL .floating-flexbox 2 assert_equals:
<div class="floating-flexbox" data-expected-width="225">
      <div style="flex: 1 1 200px; width:50px;">
        <div></div>
      </div>
      <div style="flex: 1 2 400px; width:50px;">
        <div></div>
      </div>
    </div>
width expected 225 but got 100
FAIL .floating-flexbox 3 assert_equals:
<div class="floating-flexbox" data-expected-width="225">
      <div style="flex: 1 1 200px; width:50px; min-width: 0px;">
        <div></div>
      </div>
      <div style="flex: 1 2 400px; width:50px; min-width: 0px;">
        <div></div>
      </div>
    </div>
width expected 225 but got 100
FAIL .floating-flexbox 4 assert_equals:
<div class="floating-flexbox" data-expected-width="300">
      <!-- min contribution: 100 -->
      <!-- fraction: -inf -->
      <!-- flex base size + product: 200px + -0.75*0px = 200px -->
      <div style="flex: 1 0 200px; width:50px;">
        <div></div>
      </div>
      <!-- min contribution: 100 -->
      <!-- fraction: -0.75 -->
      <!-- flex base size + product: 400px + -0.75*400px = 100px -->
      <div style="flex: 1 1 400px; width:50px;">
        <div></div>
      </div>
    </div>
width expected 300 but got 100
FAIL .floating-flexbox 5 assert_equals:
<div class="floating-flexbox" data-expected-width="200">
      <div style="flex: 0 0 50px; width: 200px;">
        <div></div>
      </div>
      <div style="flex: 0 0 50px; width: 200px;">
        <div></div>
      </div>
    </div>
width expected 200 but got 400
FAIL .floating-flexbox 6 assert_equals:
<div class="floating-flexbox" data-expected-width="600">
      <!-- contribution: 200 -->
      <!-- fraction: 150 -->
      <!-- 50 + 1*150 = 200 -->
      <div style="flex: 1 0 50px; width: 200px;">
        <div></div>
      </div>
      <!-- contribution: 200 -->
      <!-- fraction: 100 -->
      <!-- 100 + 2*150 = 400 -->
      <div style="flex: 2 0 100px; width: 200px;">
        <div></div>
      </div>
    </div>
width expected 600 but got 400
FAIL .floating-flexbox 7 assert_equals:
<div class="floating-flexbox" data-expected-width="400">
      <!-- min contribution: 100 -->
      <!-- fraction: -0.5 -->
      <!-- flex base size + product: = 200px + 50px*0 = 200px -->
      <div style="flex: 0 1 200px; width: 50px;">
        <div></div>
      </div>
      <!-- min contribution: 200 -->
      <!-- fraction: (200px - 100px) / 2 = 50px -->
      <!-- flex base size + product: = 100px + 50px*2 = 200px -->
      <div style="flex: 2 0 100px; width: 200px;">
        <div></div>
      </div>
    </div>
width expected 400 but got 250
